// final-project-client/src/app/types/user.ts
export interface User {
  username: string;
  email: string;
  password: string;
  firstName: string;
  lastName: string;
  appsCreated: string[];
  appsBought: string[];
    appsLiked: string[];
  _id: string;
  country: string;
  city: string;
  phoneNumber: string;
  aboutMe: string;
  imgUrl: string;
  __v: number;
}